Refactor Aspect Ratio to 16:9

+++ scripts/save_handler.gd
@@ -34,14 +34,9 @@ func fetch_saved_scripts():
 		saved_scripts.append("/" + node.get_path().get_concatenated_names().c_escape())
 
 func save_data(saveFolder: String = 'save') -> bool:
-	var saveFileLocation: String = get_save_file_location(saveFolder)
-	if saveFolder != 'save':
-		var success = copy_save('save', saveFolder)
-		if not success:
-			printerr('SaveHandler save_data error on copying auto-save to ', saveFolder)
-			return false
-	else:
-		create_save_subdirs(saveFileLocation)
+	# first save to the auto-save file
+	var saveFileLocation: String = get_save_file_location('save')
+	create_save_subdirs(saveFileLocation)
 	fetch_saved_scripts()
 	for script_path in saved_scripts:
 		#print(script_path)
@@ -57,6 +52,12 @@ func save_data(saveFolder: String = 'save') -> bool:
 		var err = DirAccess.remove_absolute(saveFileLocation + battle_file)
 		if err != 0:
 			printerr('SaveHandler save_data error on removing Battle file after no longer needed: error' , err)
+	# then if not only saving to the auto-save file, copy over the auto-save to the destination
+	if saveFolder != 'save':
+		var success = copy_save('save', saveFolder)
+		if not success:
+			printerr('SaveHandler save_data error on copying auto-save to ', saveFolder)
+			return false
 	return true
 
 func load_data(saveFolder: String = 'save'):
@@ -66,7 +67,7 @@ func load_data(saveFolder: String = 'save'):
 			printerr('Error clearing out auto-save to load ', saveFolder)
 			#return
 	var saveFileLocation: String = get_save_file_location('save')
-	create_save_subdirs('save')
+	create_save_subdirs(saveFileLocation)
 	fetch_saved_scripts()
 	for script_path in saved_scripts:
 		var scr = get_node_or_null(NodePath(script_path))
@@ -129,7 +130,9 @@ func copy_save(fromFolder: String, toFolder: String) -> bool:
 	var fromSaveFileLocation: String = get_save_file_location(fromFolder).trim_suffix('/')
 	var toSaveFileLocation: String = get_save_file_location(toFolder).trim_suffix('/')
 	# clear out save file destination first
-	var err = delete_directory_recursive(toSaveFileLocation)
+	var err = 0
+	if save_file_exists(toFolder):
+		err = delete_directory_recursive(toSaveFileLocation)
 	if err != 0:
 		printerr('DirAccess delete dir for copy ', toFolder, ' error ', err)
 		return false
